signal on semaphore after successful write
As of a0e09559e1e91d359d9789ee5fe909c5dcebfc16 we deliver stats to stdio
instead of whatever the usual plugin interface is. To avoid overloading
downstream databases, normal plugins have a way of being told when the
receiver of their stats has successfully committed them. Now that we're
an external plugin and don't use that interface, we don't have that
signal and so we never drain the `sem` channel. This causes the plugin
to unconditionally stall out after processing
`defaultMaxUndeliveredMessages` (1000) from the message bus and
inserting them into influxdb.

To avoid this stall, we can use the blocking behavior of `fmt.Printf()`
as a rough proxy for the old signal and drain the channel once we write
all of the influx-formatted messages to stdio.